requirements.txt
MLB-StatsAPI>=1.9.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
